requirements.txt
Flask==3.0.0
mysql-connector-python==8.3.0
bcrypt==4.1.2
python-dotenv==1.0.1
